test_stats_std for fdatairregular, add FDataGrid to skfda __init__, fix test reductions in test_irregular_operations.py, fix correction default value in stats.std and its tests
