+++ skfda/__init__.py
@@ -17,7 +17,9 @@ __getattr__, __dir__, __all__ = lazy.attach(
         "representation",
     ],
     submod_attrs={
-        'representation': ["FData", "FDataBasis", "FDataGrid"],
+        'representation': [
+            "FData", "FDataBasis", "FDataGrid", "FDataIrregular",
+        ],
         'representation._functional_data': ['concatenate'],
     },
 )
